project restructure
moved files around to follow Cobra's recommended file structure.
moved main.go to base dir and moved everything from the cmd/menu to cmd/
also updated package accordingly
